Add instructions for installing BuildStream in virtual environments
This was identified as one of the blockers in the recent discussion
about BuildStream 2.0. This patch adds instructions for installing
BuildStream stable and latest development snapshot, which are correct as
of now. However, once BuildStream 2.0 is released, we will need to
change the instructions to refer to BuildStream 2.0 as stable and
BuildStream 1.0 as the old version.